remove auctioneer runstates
stream initialization now happens as part of their
impl. separating the auctioneer state into a
starting and running phase is no longer necessary.
